add counter on cart page

+++ src/components/Product/Deatil.tsx
@@ -1,34 +1,38 @@
 import { useState } from "react";
 import { Button } from "../ui/button";
-import { useAppDispatch } from "@/redux/hooks/ReduxHook";
-import { addCart, removeCart } from "@/redux/features/Cart/cartSlice";
+import { useAppDispatch, useAppSelector } from "@/redux/hooks/ReduxHook";
+import { addCart, cartData, removeCart } from "@/redux/features/Cart/cartSlice";
 import { toast } from "sonner";
-import { removeCount, totalCount } from "@/redux/features/Cart/totalPriceSlice";
+import Counter from "../counter/Counter";
+import { TProduct } from "@/types/TProducts";
 
-const Deatil = ({ data }: { data: Record<string, string> }) => {
-  const dispatch = useAppDispatch();
-  const [quantity, setQuantity] = useState(1);
+const Deatil = ({ data }: { data: TProduct }) => {
   const { _id, name, price, stock, description, image, category } = data;
+  const dispatch = useAppDispatch();
+  const cart = useAppSelector(cartData);
+  const existQuantity = cart?.find((item) => item.productId === _id);
+  const [quantity, setQuantity] = useState(existQuantity?.quantity || 1);
   const handleClick = () => {
     const cartData = {
       productId: _id,
-      quantity,
+      quantity: quantity,
+      price,
     };
-    const total = Number(price) * quantity;
-    dispatch(totalCount(total));
+    if (existQuantity?.quantity === quantity) {
+      toast.warning("Please Add or Remove item Produte ");
+      return;
+    }
     dispatch(addCart(cartData));
 
     toast("Product is add in your cart", {
       action: {
         label: "Undo",
         onClick: () => {
-          dispatch(removeCount(total));
           dispatch(removeCart(_id));
         },
       },
     });
   };
-
   return (
     <div>
       <h2 className="capitalize font-bold text-3xl text-center text-textColor">
@@ -44,7 +48,10 @@ const Deatil = ({ data }: { data: Record<string, string> }) => {
               <span className="text-textColor font-semibold">name:</span> {name}
             </h3>
             <h4 className="capitalize text-textColor text-xl font-semibold">
-              ${price}
+              $
+              {existQuantity?.quantity
+                ? price * existQuantity?.quantity
+                : price}
             </h4>
           </div>
           <div className=" flex flex-wrap justify-between">
@@ -58,28 +65,21 @@ const Deatil = ({ data }: { data: Record<string, string> }) => {
               <span className="capitalize text-textColor text-xl font-semibold">
                 Stock:
               </span>{" "}
-              {stock}
+              {existQuantity?.quantity
+                ? stock - existQuantity?.quantity
+                : stock}
             </h4>
           </div>
           <div className=" flex flex-wrap justify-between">
-            <div className="flex rounded-lg">
-              <button
-                onClick={() => quantity > 1 && setQuantity(quantity - 1)}
-                className="text-xl text-textColor font-bold border-textColor border py-1 px-4 rounded-l-lg hover:bg-textColor hover:text-white"
-              >
-                -
-              </button>
-              <p className="text-xl text-textColor font-bold border-textColor border-y py-1 px-4">
-                {quantity}
-              </p>
-              <button
-                onClick={() => quantity < 30 && setQuantity(quantity + 1)}
-                className="text-xl text-textColor font-bold border-textColor border py-1 px-4 rounded-r-lg hover:bg-textColor hover:text-white"
-              >
-                +
-              </button>
-            </div>
-            <Button onClick={handleClick}>Add Cart</Button>
+            {existQuantity?.productId ? (
+              <Counter
+                quantity={quantity}
+                setQuantity={setQuantity}
+                data={data}
+              />
+            ) : (
+              <Button onClick={handleClick}>Add Cart</Button>
+            )}
           </div>
           <hr />
           <p>
